Update 주유소.java
입력 받을 때는 int 가능하다 (10억 이하)
비용 구할 때는 long으로 받아야 한다. (거리 * 비용)

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -8,27 +8,27 @@ public class Main {
 
         int n = Integer.parseInt(br.readLine());
 
-        long[] distance = new long[n - 1];
+        int[] distance = new int[n - 1];
         String[] input = br.readLine().split(" ");
         for (int i = 0; i < n - 1; i++)
-            distance[i] = Long.parseLong(input[i]);
+            distance[i] = Integer.parseInt(input[i]);
 
-        long[] cost = new long[n];
+        int[] cost = new int[n];
         input = br.readLine().split(" ");
         for (int i = 0; i < n; i++)
-            cost[i] = Long.parseLong(input[i]);
+            cost[i] = Integer.parseInt(input[i]);
 
-        long min = Integer.MAX_VALUE;
+        int min = Integer.MAX_VALUE;
         long sum = 0;
         for (int i = 0; i < n - 1; i++) {
             min = Math.min(min, cost[i]);
-            long nextDistance = distance[i];
+            int nextDistance = distance[i];
             sum += getCost(nextDistance, min);
         }
         System.out.println(sum);
     }
 
-    private static long getCost(long distance, long cost) {
-        return distance * cost;
+    private static long getCost(int distance, int cost) {
+        return (long) distance * cost;
     }
 }
